Refactor: move all relocation related items to separate module
