+++ src/state_machine.rs
@@ -11,10 +11,10 @@ use crate::{
     chain::{EldersInfo, GenesisPfxInfo},
     error::RoutingError,
     id::{P2pNode, PublicId},
-    messages::{RelocatePayload, SignedRelocateDetails},
     network_service::NetworkBuilder,
     outbox::EventBox,
     pause::PausedState,
+    relocation::{RelocatePayload, SignedRelocateDetails},
     routing_table::Prefix,
     states::common::Base,
     states::{Adult, BootstrappingPeer, Elder, JoiningPeer},
