+++ src/states/adult.rs
@@ -20,12 +20,12 @@ use crate::{
     event::Event,
     id::{FullId, P2pNode, PublicId},
     messages::{
-        BootstrapResponse, DirectMessage, HopMessage, RelocateDetails, RoutingMessage,
-        SignedRoutingMessage,
+        BootstrapResponse, DirectMessage, HopMessage, RoutingMessage, SignedRoutingMessage,
     },
     outbox::EventBox,
     parsec::{DkgResultWrapper, ParsecMap},
     peer_map::PeerMap,
+    relocation::RelocateDetails,
     routing_message_filter::RoutingMessageFilter,
     routing_table::{Authority, Prefix},
     state_machine::{State, Transition},
