+++ src/states/bootstrapping_peer.rs
@@ -15,12 +15,10 @@ use crate::{
     error::{InterfaceError, RoutingError},
     event::Event,
     id::{FullId, P2pNode},
-    messages::{
-        BootstrapResponse, DirectMessage, HopMessage, RelocatePayload, RoutingMessage,
-        SignedRelocateDetails,
-    },
+    messages::{BootstrapResponse, DirectMessage, HopMessage, RoutingMessage},
     outbox::EventBox,
     peer_map::PeerMap,
+    relocation::{RelocatePayload, SignedRelocateDetails},
     routing_table::{Authority, Prefix},
     state_machine::{State, Transition},
     states::JoiningPeer,
