+++ src/states/common/approved.rs
@@ -15,9 +15,10 @@ use crate::{
     error::RoutingError,
     event::Event,
     id::PublicId,
-    messages::{DirectMessage, MessageContent, RelocateDetails, RoutingMessage},
+    messages::{DirectMessage, MessageContent, RoutingMessage},
     outbox::EventBox,
     parsec::{self, Block, DkgResultWrapper, Observation, ParsecMap},
+    relocation::RelocateDetails,
     routing_table::{Authority, Prefix},
     state_machine::Transition,
     types::MessageId,
