+++ src/states/joining_peer.rs
@@ -15,12 +15,10 @@ use crate::{
     chain::{GenesisPfxInfo, NetworkParams},
     error::{InterfaceError, RoutingError},
     id::{FullId, P2pNode},
-    messages::{
-        DirectMessage, HopMessage, MessageContent, RelocatePayload, RoutingMessage,
-        SignedRoutingMessage,
-    },
+    messages::{DirectMessage, HopMessage, MessageContent, RoutingMessage, SignedRoutingMessage},
     outbox::EventBox,
     peer_map::PeerMap,
+    relocation::RelocatePayload,
     routing_message_filter::RoutingMessageFilter,
     routing_table::Authority,
     state_machine::{State, Transition},
